+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @equaltoai/playground
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @equaltoai/greater-components-fediverse@1.0.1
+
 ## 0.0.2
 
 ### Patch Changes